requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask~=2.2.3
fonttools==4.38.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
Werkzeug==2.2.2

pyserial~=3.5